Support busy-looping for small phase resolution
The closest time between interrupts is about 3us, so rework the waveform
generator to support phases down to 1us, while limiting the time in the
IRQ handler to avoid WDT or WiFi errors.

Waveforms now are based off of absolute cycle, not delta-cycles, and
the GPIOs are converted from a bit to a mask (to save clocks in the IRQ)
costing about 40 more bytes in global storage and reducing the minimum
period down from 3.5us to 1.20us at 160MHz.

Steppers still based off of delta-times.

+++ cores/esp8266/core_esp8266_waveform.h
@@ -64,9 +64,9 @@ extern "C" {
 int startWaveform(uint8_t pin, uint32_t timeHighUS, uint32_t timeLowUS, uint32_t runTimeUS);
 int stopWaveform(uint8_t pin);
 
-int setStepperDirPin(int pin);
-int pushStepperMove(int pin, int dir, int sync, uint16_t pulses, float j, float a0, float v0);
-int removeStepper(int pin);
+int setStepperDirPin(uint8_t pin);
+int pushStepperMove(uint8_t pin, int dir, int sync, uint16_t pulses, float j, float a0, float v0);
+int removeStepper(uint8_t pin);
 
 #ifdef __cplusplus
 }
